Split shell connection lease tests
